components/context-sync/requirements.txt
requests>=2.25.0
python-dotenv>=0.19.0
html2text>=2020.1.16
